Revert "Added Windows support to the script""
